back-end/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
Flask==2.1.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
nulltype==2.3.1
oauthlib==3.2.2
plaid-python==14.0.0
psycopg==3.1.10
psycopg-binary==3.1.10
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
